backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
APScheduler==3.10.4
argcomplete==3.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
bitarray==2.9.2
cachetools==5.3.3
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
ckzg==1.0.2
click==8.1.7
colorama==0.4.6
cytoolz==0.12.3
distlib==0.3.8
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.3.1
eth-utils==4.1.1
eth_abi==5.1.0
filelock==3.15.4
flake8==7.1.0
frozenlist==1.4.1
hexbytes==0.3.1
idna==3.7
isort==5.13.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
mccabe==0.7.0
multidict==6.0.5
mypy==1.10.1
mypy-extensions==1.0.0
packaging==24.1
parsimonious==0.10.0
pipx==1.6.0
platformdirs==4.2.2
pluggy==1.5.0
protobuf==5.27.2
pycodestyle==2.12.0
pycryptodome==3.20.0
pyflakes==3.2.0
PyJWT==2.8.0
pyproject-api==1.7.1
python-dotenv==1.0.1
pytz==2024.1
pyunormalize==15.1.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rest-framework-simplejwt==0.0.2
rlp==4.0.1
rpds-py==0.18.1
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
toolz==0.12.1
tox==4.16.0
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
userpath==1.9.2
virtualenv==20.26.3
web3==6.20.0
websockets==12.0
yarl==1.9.4
